Add Shop data loader command.
This is very simple. A CSV file that gets parsed and imported using the Shop model.

There are categories that have multiple entries in the row. That will make selection/comparison
more complicated so I will need to think about how I want to do that.

For such a small non serious project checking for a comma in that column and then splitting on it is
feasible.

Could just enforce a single category per entry.

Anyways the Shop data is now loaded into the sqlite DB and it can be modified through the Admin interface.

The command to load CSV data is python manage.py load_shop_data

As of now if the Shop data already exists it just exists so I will be changing that in the future
to allow loading of new/changed data.